+++ internal/ent/cabinet/cabinet.go
@@ -132,7 +132,7 @@ var (
 	// UpdateDefaultUpdatedAt holds the default value on update for the "updated_at" field.
 	UpdateDefaultUpdatedAt func() time.Time
 	// DefaultHealth holds the default value on creation for the "health" field.
-	DefaultHealth uint
+	DefaultHealth uint8
 	// DefaultBatteryNum holds the default value on creation for the "battery_num" field.
 	DefaultBatteryNum uint
 	// DefaultBatteryFullNum holds the default value on creation for the "battery_full_num" field.
